Major update and rework of most pages
First, added Gemfile to preview website locally, see updated README for instructions

Second, major rework of index.md was performed. This was to improve its styles and add new features, such as a banner, short bio with stats and featured writeups.

For this last element, many changes were made. The "Hot writeups" button on the navbar was redundant, therefore it was removed. Then, in order to sort writeups chronologically, each writeup is now tagged with a title, date and banner path for images.

Speaking of images, every external link was removed in favor of locally stored assets (this excludes THM's iframe for stats)

Lastly, minor fixes were made to writeups for consistency and favicon was fixed.